剑指 Offer 41. 数据流中的中位数

+++ leetcode/src/main/java/com/leetcode/tag/must1/seven/MedianFinder1.java
@@ -36,9 +36,9 @@ public class MedianFinder1 {
 
     public double findMedian() {
         if (count % 2 == 0) {
-            return (maxheap.peek() + minheap.peek()) / 2;
+            return (double) (maxheap.peek() + minheap.peek()) / 2;
         }
 
-        return maxheap.peek();
+        return (double) maxheap.peek();
     }
 }
